pyproject.toml
[project]
name = "tds-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "duckdb>=1.4.2",
    "fastapi>=0.120.0",
    "faster-whisper>=1.2.1",
    "geopandas>=1.1.1",
    "geopy>=2.4.1",
    "ijson>=3.4.0.post0",
    "markitdown>=0.1.3",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pymupdf4llm>=0.0.27",
    "pytest-playwright>=0.7.1",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "tabula-py>=2.10.0",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
    "yt-dlp>=2025.10.22",
]
